% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-data-frame-har.R
\name{as_data_frame.harentry}
\alias{as_data_frame.harentry}
\alias{as_data_frame.harentries}
\alias{as_data_frame.har}
\alias{as.data.frame.har}
\alias{as.data.frame.harentries}
\alias{as.data.frame.harentry}
\title{Turns a "HAR"-like object into a data frame(tibble)}
\usage{
as_data_frame.harentry(x, ...)

as_data_frame.harentries(x, ...)

as_data_frame.har(x, ...)

\method{as.data.frame}{har}(x, ...)

\method{as.data.frame}{harentries}(x, ...)

\method{as.data.frame}{harentry}(x, ...)
}
\arguments{
\item{x}{A \code{harentry} object}

\item{...}{ignored}
}
\value{
data frame (tibble)
}
\description{
Turns a "HAR"-like object into a data frame(tibble)
}
