% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-har.r
\name{as_har}
\alias{as_har}
\title{Turn a generic Splash HAR response into a HAR object}
\usage{
as_har(splash_resp)
}
\arguments{
\item{as_har}{splash response object as returned by `splash:har()` lua script}
}
\description{
Turn a generic Splash HAR response into a HAR object
}
