% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_req.R
\name{as_httr_req}
\alias{as_httr_req}
\title{Create an httr verb request function from an HAR request}
\usage{
as_httr_req(entry, quiet = TRUE)
}
\arguments{
\item{entry}{HAR entry}

\item{quiet}{quiet (no messages)}
}
\description{
This function is very useful if you used \code{splashr} to find XHR requests in a dynamic
page and want to be able to make a call directly to that XHR resource. Once you
identify the proper HAR entry, pass it to this function and fully working function
that makes an \code{httr::VERB()} request will be created and returned.
}
