% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get_content_size}
\alias{get_content_size}
\alias{get_body_size}
\alias{get_headers_size}
\title{Retrieve size of content |  body | headers}
\usage{
get_content_size(har_resp_obj)

get_body_size(har_resp_obj)

get_headers_size(har_resp_obj)
}
\arguments{
\item{har_resp_obj}{HAR response object}
}
\description{
Retrieve size of content |  body | headers
}
