% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_content_type}
\alias{get_content_type}
\alias{is_content_type}
\alias{is_json}
\alias{is_xml}
\alias{is_css}
\alias{is_plain}
\alias{is_binary}
\alias{is_javascript}
\alias{is_html}
\alias{is_jpeg}
\alias{is_png}
\alias{is_svg}
\alias{is_gif}
\alias{is_xhr}
\title{Retrieve or test content type of a HAR request object}
\usage{
get_content_type(har_resp_obj)

is_content_type(har_resp_obj, type = "application/json")

is_json(har_resp_obj)

is_xml(har_resp_obj)

is_css(har_resp_obj)

is_plain(har_resp_obj)

is_binary(har_resp_obj)

is_javascript(har_resp_obj)

is_html(har_resp_obj)

is_jpeg(har_resp_obj)

is_png(har_resp_obj)

is_svg(har_resp_obj)

is_gif(har_resp_obj)

is_xhr(har_resp_obj)
}
\arguments{
\item{har_resp_obj}{a reponse object from [render_har()] or [execute_lua()]}

\item{type}{content type to compare to (default: "\code{application/json}")}
}
\description{
Retrieve or test content type of a HAR request object
}
\seealso{
Other splash_har_helpers: \code{\link{get_content_size}},
  \code{\link{get_har_entry}},
  \code{\link{get_request_type}},
  \code{\link{get_request_url}},
  \code{\link{get_response_body}},
  \code{\link{har_entry_count}}
}
\concept{splash_har_helpers}
