% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get_har_entry}
\alias{get_har_entry}
\title{Retrieve an entry by index from a HAR object}
\usage{
get_har_entry(x, i = 1)
}
\arguments{
\item{x}{can be a `har` object, `harlog` object or `harentries` object}

\item{i}{index of the HAR entry to retrieve}
}
\description{
Retrieve an entry by index from a HAR object
}
