% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_headers}
\alias{get_headers}
\title{Retrieve response headers as a data frame}
\usage{
get_headers(har_resp_obj)
}
\arguments{
\item{har_resp_obj}{HAR response object}
}
\description{
Retrieve response headers as a data frame
}
\note{
the \code{name} column that contains the header key is normalized to lower case
}
\seealso{
Other splash_har_helpers: \code{\link{get_content_size}},
  \code{\link{get_content_type}},
  \code{\link{get_har_entry}},
  \code{\link{get_header_val}},
  \code{\link{get_request_type}},
  \code{\link{get_request_url}},
  \code{\link{get_response_body}},
  \code{\link{get_response_url}},
  \code{\link{har_entry_count}}
}
\concept{splash_har_helpers}
