% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get_response_body}
\alias{get_response_body}
\title{Retrieve the body content of a HAR entry}
\usage{
get_response_body(har_resp_obj)
}
\arguments{
\item{har_resp_obj}{HAR response object}
}
\value{
A \code{raw} vector of the content or \code{NULL}
}
\description{
Retrieve the body content of a HAR entry
}
