% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{har_entries}
\alias{har_entries}
\title{Retrieve just the HAR entries from a splashr request}
\usage{
har_entries(x)
}
\arguments{
\item{x}{can be a `har` object, `harlog` object or `harentries` object}
}
\description{
Retrieve just the HAR entries from a splashr request
}
