% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{har_entry_count}
\alias{har_entry_count}
\title{Retrieves number of HAR entries in a response}
\usage{
har_entry_count(x)
}
\arguments{
\item{x}{can be a `har` object, `harlog` object or `harentries` object}
}
\description{
Retrieves number of HAR entries in a response
}
\seealso{
Other splash_har_helpers: \code{\link{get_content_size}},
  \code{\link{get_content_type}},
  \code{\link{get_har_entry}},
  \code{\link{get_header_val}}, \code{\link{get_headers}},
  \code{\link{get_request_type}},
  \code{\link{get_request_url}},
  \code{\link{get_response_body}},
  \code{\link{get_response_url}}
}
\concept{splash_har_helpers}
