% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker-splash.r, R/docker.r
\name{install_splash}
\alias{install_splash}
\alias{install_splash}
\title{Retrieve the Docker image for Splash}
\usage{
install_splash(host = harbor::localhost)

install_splash(host = harbor::localhost)
}
\arguments{
\item{host}{Docker host; defaults to \code{localhost}}

\item{host}{Docker host; defauolts to \code{localhost}}
}
\value{
\code{harbor} \code{host} object

\code{harbor} \code{host} object
}
\description{
Retrieve the Docker image for Splash

Retrieve the Docker image for Splash
}
\examples{
\dontrun{
install_splash()
splash_container <- start_splash()
stop_splash(splash_container)
}
\dontrun{
install_splash()
splash_container <- start_splash()
stop_splash(splash_container)
}
}
