% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{json_fromb64}
\alias{json_fromb64}
\title{Convert a Base64 encoded string into an R object}
\usage{
json_fromb64(x, flatten = TRUE, ...)
}
\arguments{
\item{x}{a string}

\item{flatten}{flatten JSON structures upon conversion?}

\item{...}{passed on to \code{jsonlite::fromJSON()}}
}
\description{
A simple wrapper around calls to \code{openssl::base64_decode()} and
\code{jsonlite::fromJSON()}.
}
