% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker-splash.R
\name{killall_splash}
\alias{killall_splash}
\title{Prune all dead and running Splash Docker containers}
\usage{
killall_splash()
}
\description{
_This is a destructive function._ It will stop **any** Docker container that
is based on an image matching "`scrapinghub/splashr`". It's best used when you
had a session forcefully interuppted and had been using the R helper functions
to start/stop the Splash Docker container. You may want to consider using the
Docker command-line interface to perform this work manually.
}
