% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splashr.r
\name{print.splashr}
\alias{print.splashr}
\title{Print for splashr}
\usage{
\method{print}{splashr}(x, ...)
}
\arguments{
\item{x}{a splashr object}

\item{...}{unused}
}
\value{
x
}
\description{
Print for splashr
}
\keyword{internal}
