% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_file.R
\name{render_file}
\alias{render_file}
\title{Return the HTML or image (png) of the javascript-rendered page in a local file}
\usage{
render_file(splash_obj, file_path, output = c("html", "png"), wait = 0,
  viewport = "1024x768", ...)
}
\arguments{
\item{splash_obj}{Object created by a call to \url{splash}}

\item{file_path}{Absolute path to a filename on the local host. \strong{This only works with a locally running Splash instance started with \url{start_splash}.}}

\item{output}{either \code{html} or \code{png} to get the page content or an image capture}

\item{viewport}{View width and height (in pixels) of the browser viewport to render the web page. Format is "\code{<width>x<height>}". e.g. 800x600. Default value is 1024x768.}

\item{...}{other params to \url{render_html} or \url{render_png}}
}
\value{
An XML document or \code{magick} object
}
\description{
The suggested use-case for this is rendering a widget
}
\details{
TODO Test if container is running
TODO Enable passing in of an htmlwidget and use saveWidget
}
