% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-har.r
\name{render_har}
\alias{render_har}
\title{Return information about Splash interaction with a website in HAR format.}
\usage{
render_har(splash_obj = splash_local, url, base_url, response_body = FALSE,
  timeout = 30, resource_timeout, wait = 0, proxy, js, js_src, filters,
  allowed_domains, allowed_content_types, forbidden_content_types,
  viewport = "1024x768", images, headers, body, http_method, save_args,
  load_args)
}
\arguments{
\item{splash_obj}{Object created by a call to \code{\link[=splash]{splash()}}}

\item{url}{The URL to render (required)}

\item{base_url}{The base url to render the page with.}

\item{response_body}{When \code{TRUE}, response content is included in the HAR records}

\item{timeout}{A timeout (in seconds) for the render (defaults to 30).}

\item{resource_timeout}{A timeout (in seconds) for individual network requests.}

\item{wait}{Time (in seconds) to wait for updates after page is loaded (defaults to 0).}

\item{proxy}{Proxy profile name or proxy URL.}

\item{js}{Javascript profile name.}

\item{js_src}{JavaScript code to be executed in page context.}

\item{filters}{Comma-separated list of request filter names.}

\item{allowed_domains}{Comma-separated list of allowed domain names. If present, Splash won’t load anything neither from domains not in this list nor from subdomains of domains not in this list.}

\item{allowed_content_types}{Comma-separated list of allowed content types. If present, Splash will abort any request if the response’s content type doesn’t match any of the content types in this list. Wildcards are supported.}

\item{forbidden_content_types}{Comma-separated list of forbidden content types. If present, Splash will abort any request if the response’s content type matches any of the content types in this list. Wildcards are supported.}

\item{viewport}{View width and height (in pixels) of the browser viewport to render the web page. Format is “<width>x<height>”, e.g. 800x600. Default value is 1024x768.}

\item{images}{Whether to download images.}

\item{headers}{HTTP headers to set for the first outgoing request.}

\item{body}{Body of HTTP POST request to be sent if method is POST.}

\item{http_method}{HTTP method of outgoing Splash request.}

\item{save_args}{A list of argument names to put in cache.}

\item{load_args}{Parameter values to load from cache}
}
\value{
a \link{HARtools} \code{har} object
}
\description{
It includes information about requests made, responses received, timings, headers, etc and
is incredibly detailed, full of information on every component loaded.
}
\references{
\href{http://splash.readthedocs.io/en/stable/index.html}{Splash docs}
}
