% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-html.r
\name{render_html}
\alias{render_html}
\title{Return the HTML of the javascript-rendered page.}
\usage{
render_html(splash_obj, url, base_url, timeout = 30,
  resource_timeout = NULL, wait = 0, proxy, js, js_src, filters,
  allowed_domains = "", allowed_content_types = "",
  forbidden_content_types = "", viewport = "1024x768", images, headers,
  body, http_method, save_args, load_args)
}
\arguments{
\item{splash_obj}{Object created by a call to \link{splash}}

\item{url}{The URL to render (required)}

\item{base_url}{TBD The base url to render the page with.}

\item{timeout}{TBD A timeout (in seconds) for the render (defaults to 30).}

\item{resource_timeout}{A timeout (in seconds) for individual network requests.}

\item{wait}{Time (in seconds) to wait for updates after page is loaded (defaults to 0).}

\item{proxy}{TBD Proxy profile name or proxy URL.}

\item{js}{TBD Javascript profile name.}

\item{js_src}{TBD JavaScript code to be executed in page context.}

\item{filters}{TBD Comma-separated list of request filter names.}

\item{allowed_domains}{TBD Comma-separated list of allowed domain names. If present, Splash won’t load anything neither from domains not in this list nor from subdomains of domains not in this list.}

\item{allowed_content_types}{TBD Comma-separated list of allowed content types. If present, Splash will abort any request if the response’s content type doesn’t match any of the content types in this list. Wildcards are supported.}

\item{forbidden_content_types}{TBD Comma-separated list of forbidden content types. If present, Splash will abort any request if the response’s content type matches any of the content types in this list. Wildcards are supported.}

\item{viewport}{View width and height (in pixels) of the browser viewport to render the web page. Format is “<width>x<height>”, e.g. 800x600. Default value is 1024x768.}

\item{images}{TBD Whether to download images.}

\item{headers}{TBD HTTP headers to set for the first outgoing request.}

\item{body}{TBD Body of HTTP POST request to be sent if method is POST.}

\item{http_method}{TBD HTTP method of outgoing Splash request.}

\item{save_args}{TBD A list of argument names to put in cache.}

\item{load_args}{TBD Parameter values to load from cache}
}
\description{
Similar to \code{rvest::read_html}.
}
