% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.r
\name{splash_click}
\alias{splash_click}
\title{Trigger mouse click event in web page.}
\usage{
splash_click(splash_obj, x, y)
}
\arguments{
\item{splash_obj}{splashr object}

\item{x, y}{coordinates (distances from the left or top, relative to the current viewport)}
}
\description{
Trigger mouse click event in web page.
}
