% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{splash_enable_javascript}
\alias{splash_enable_javascript}
\title{Enable or disable execution of JavaSript code embedded in the page.}
\usage{
splash_enable_javascript(splash_obj, enable = TRUE)
}
\arguments{
\item{splash_obj}{splashr object}

\item{enable}{logical}
}
\description{
JavaScript execution is enabled by default.
}
\examples{
\dontrun{
splash_local \%>\%
  splash_response_body(TRUE) \%>\%
  splash_private_mode(TRUE) \%>\%
  splash_enable_javascript(FALSE) \%>\%
  splash_user_agent(ua_macos_chrome) \%>\%
  splash_go("https://rud.is/b") \%>\%
  splash_wait(2) \%>\%
  splash_har() -> rud_har
}
}
\seealso{
Other splash_dsl_attributes: 
\code{\link{splash_images}()},
\code{\link{splash_plugins}()},
\code{\link{splash_private_mode}()},
\code{\link{splash_response_body}()}
}
\concept{splash_dsl_attributes}
