% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{splash_focus}
\alias{splash_focus}
\title{Focus on a document element provided by a CSS selector}
\usage{
splash_focus(splash_obj, selector)
}
\arguments{
\item{splash_obj}{splashr object}

\item{selector}{valid CSS selector}
}
\description{
Focus on a document element provided by a CSS selector
}
\references{
See \href{https://splash.readthedocs.io/en/stable/scripting-ref.html#splash-send-text}{the docs} for more info
}
\seealso{
Other splash_dsl_functions: 
\code{\link{splash_add_lua}()},
\code{\link{splash_click}()},
\code{\link{splash_go}()},
\code{\link{splash_har_reset}()},
\code{\link{splash_har}()},
\code{\link{splash_html}()},
\code{\link{splash_png}()},
\code{\link{splash_press}()},
\code{\link{splash_release}()},
\code{\link{splash_send_keys}()},
\code{\link{splash_send_text}()},
\code{\link{splash_wait}()}
}
\concept{splash_dsl_functions}
