% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.r
\name{splash_go}
\alias{splash_go}
\title{Go to an URL.}
\usage{
splash_go(splash_obj, url)
}
\arguments{
\item{splash_obj}{splashr object}

\item{url}{- URL to load;}
}
\description{
This is similar to entering an URL in a browser address bar, pressing Enter and waiting
until page loads.
}
\examples{
\dontrun{
splash_local \%>\%
  splash_response_body(TRUE) \%>\%
  splash_user_agent(ua_macos_chrome) \%>\%
  splash_go("https://rud.is/b") \%>\%
  splash_wait(2) \%>\%
  splash_har() -> rud_har
}
}
