% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{splash_har_reset}
\alias{splash_har_reset}
\title{Drops all internally stored HAR records.}
\usage{
splash_har_reset(splash_obj)
}
\arguments{
\item{splash_obj}{splashr object}
}
\description{
Drops all internally stored HAR records.
}
\seealso{
Other splash_dsl_functions: 
\code{\link{splash_add_lua}()},
\code{\link{splash_click}()},
\code{\link{splash_focus}()},
\code{\link{splash_go}()},
\code{\link{splash_har}()},
\code{\link{splash_html}()},
\code{\link{splash_png}()},
\code{\link{splash_press}()},
\code{\link{splash_release}()},
\code{\link{splash_send_keys}()},
\code{\link{splash_send_text}()},
\code{\link{splash_wait}()}
}
\concept{splash_dsl_functions}
