% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splashr.r
\name{splash_history}
\alias{splash_history}
\title{Get information about requests/responses for the pages loaded}
\usage{
splash_history(splash_obj = splash_local)
}
\arguments{
\item{splash_obj}{A splash connection object}
}
\description{
Get information about requests/responses for the pages loaded
}
\examples{
\dontrun{
sp <- splash()
splash_history(sp)
}
}
\seealso{
Other splash_info_functions: \code{\link{splash_active}},
  \code{\link{splash_debug}},
  \code{\link{splash_perf_stats}},
  \code{\link{splash_version}}
}
