% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.r
\name{splash_html}
\alias{splash_html}
\title{Return a HTML snapshot of a current page.}
\usage{
splash_html(splash_obj, raw_html = FALSE)
}
\arguments{
\item{splash_obj}{splashr object}

\item{raw_html}{if \code{TRUE} then return a character vector vs an XML document.}
}
\description{
Similar to \link{render_html} but used in a script context. Should be the LAST element in
a DSL script chain as this will execute the script and return the HTML content
}
\examples{
\dontrun{
splash_local \%>\%
  splash_response_body(TRUE) \%>\%
  splash_user_agent(ua_macos_chrome) \%>\%
  splash_go("https://rud.is/b") \%>\%
  splash_wait(2) \%>\%
  splash_html() -> rud_pg
}
}
