% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splashr.r
\name{splash_perf_stats}
\alias{splash_perf_stats}
\title{Get Splash performance-related statistics}
\usage{
splash_perf_stats(splash_obj = splash_local)
}
\arguments{
\item{splash_obj}{A splash connection object}
}
\description{
Get Splash performance-related statistics
}
\examples{
\dontrun{
sp <- splash()
splash_perf_stats(sp)
}
}
\seealso{
Other splash_info_functions: \code{\link{splash_active}},
  \code{\link{splash_debug}}, \code{\link{splash_history}},
  \code{\link{splash_version}}
}
\concept{splash_info_functions}
