% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.r
\name{splash_plugins}
\alias{splash_plugins}
\title{Enable or disable browser plugins (e.g. Flash).}
\usage{
splash_plugins(splash_obj, enable = FALSE)
}
\arguments{
\item{splash_obj}{splashr object}

\item{enable}{logical}
}
\description{
Plugins are disabled by default.
}
\examples{
\dontrun{
splash_local \%>\%
  splash_plugins(TRUE) \%>\%
  splash_user_agent(ua_macos_chrome) \%>\%
  splash_go("https://rud.is/b") \%>\%
  splash_wait(2) \%>\%
  splash_har() -> rud_har
}
}
