% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.r
\name{splash_private_mode}
\alias{splash_private_mode}
\title{Enable or disable execution of JavaSript code embedded in the page.}
\usage{
splash_private_mode(splash_obj, enable = FALSE)
}
\arguments{
\item{splash_obj}{splashr object}

\item{enable}{logical}
}
\description{
Private mode is enabled by default unless you pass flag \code{--disable-private-mode}
at Splash (server) startup. Note that if you disable private mode browsing data such
as cookies or items kept in local storage may persist between requests.
}
\examples{
\dontrun{
splash_local \%>\%
  splash_response_body(TRUE) \%>\%
  splash_private_mode(TRUE) \%>\%
  splash_user_agent(ua_macos_chrome) \%>\%
  splash_go("https://rud.is/b") \%>\%
  splash_wait(2) \%>\%
  splash_har() -> rud_har
}
}
\seealso{
Other splash_dsl_attributes: \code{\link{splash_enable_javascript}},
  \code{\link{splash_images}},
  \code{\link{splash_plugins}},
  \code{\link{splash_response_body}}
}
