% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.r
\name{splash_release}
\alias{splash_release}
\title{Trigger mouse release event in web page.}
\usage{
splash_release(splash_obj, x, y)
}
\arguments{
\item{splash_obj}{splashr object}

\item{x, y}{coordinates (distances from the left or top, relative to the current viewport)}
}
\description{
Trigger mouse release event in web page.
}
