% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.r
\name{splash_response_body}
\alias{splash_response_body}
\title{Enable or disable response content tracking.}
\usage{
splash_response_body(splash_obj, enable = FALSE)
}
\arguments{
\item{splash_obj}{splashr object}

\item{enable}{logical}
}
\description{
By default Splash doesn’t keep bodies of each response in memory, for efficiency reasons.
}
\examples{
\dontrun{
splash_local \%>\%
  splash_response_body(TRUE) \%>\%
  splash_user_agent(ua_macos_chrome) \%>\%
  splash_go("https://rud.is/b") \%>\%
  splash_wait(2) \%>\%
  splash_har() -> rud_har
}
}
\seealso{
Other splash_dsl_attributes: \code{\link{splash_enable_javascript}},
  \code{\link{splash_images}},
  \code{\link{splash_plugins}},
  \code{\link{splash_private_mode}}
}
\concept{splash_dsl_attributes}
