% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{splash_send_text}
\alias{splash_send_text}
\title{Send text as input to page context, literally, character by character.}
\usage{
splash_send_text(splash_obj, text)
}
\arguments{
\item{splash_obj}{splashr object}

\item{text}{string to send}
}
\description{
This is different from \code{\link[=splash_send_keys]{splash_send_keys()}}
}
\note{
This adds a call to \code{splash:wait} so you do not have to
}
\references{
See \href{https://splash.readthedocs.io/en/stable/scripting-ref.html#splash-send-keys}{the docs} for more info
}
\seealso{
Other splash_dsl_functions: \code{\link{splash_add_lua}},
  \code{\link{splash_click}}, \code{\link{splash_focus}},
  \code{\link{splash_go}}, \code{\link{splash_har_reset}},
  \code{\link{splash_har}}, \code{\link{splash_html}},
  \code{\link{splash_png}}, \code{\link{splash_press}},
  \code{\link{splash_release}},
  \code{\link{splash_send_keys}}, \code{\link{splash_wait}}
}
\concept{splash_dsl_functions}
