% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{splash_themes}
\alias{splash_themes}
\title{List available Splash themes}
\usage{
splash_themes()
}
\value{
A character vector of registered theme names, sorted alphabetically.
}
\description{
List available Splash themes
}
\examples{
splash_themes()

}
