% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splashr.R
\name{splash_version}
\alias{splash_version}
\title{Get Splash version information}
\usage{
splash_version(splash_obj = splash_local)
}
\arguments{
\item{splash_obj}{A splash connection object}
}
\description{
Get Splash version information
}
\examples{
\dontrun{
sp <- splash()
splash_version(sp)
}
}
\seealso{
Other splash_info_functions: \code{\link{splash_active}},
  \code{\link{splash_debug}}, \code{\link{splash_history}},
  \code{\link{splash_perf_stats}}
}
\concept{splash_info_functions}
