% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.r
\name{splash_wait}
\alias{splash_wait}
\title{Wait for a period time}
\usage{
splash_wait(splash_obj, time = 2)
}
\arguments{
\item{splash_obj}{splashr object}

\item{time}{number of seconds to wait}
}
\description{
When script is waiting WebKit continues processing the webpage
}
\examples{
\dontrun{
splash_local \%>\%
  splash_response_body(TRUE) \%>\%
  splash_user_agent(ua_macos_chrome) \%>\%
  splash_go("https://rud.is/b") \%>\%
  splash_wait(2) \%>\%
  splash_har() -> rud_har
}
}
