% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splashr-package.R
\name{splashr-exports}
\alias{splashr-exports}
\alias{\%>\%}
\alias{writeHAR}
\alias{HARviewer}
\alias{renderHARviewer}
\alias{HARviewerOutput}
\title{splashr exported operators}
\description{
The following functions are imported and then re-exported
from the splashr package to enable use of the magrittr
pipe operator with no additional library calls
}
