% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker-splash.R
\name{stop_splash}
\alias{stop_splash}
\title{Stop a running a Splash server Docker container}
\usage{
stop_splash(splash_container)
}
\arguments{
\item{splash_container}{Docker `container` object created by [start_splash()]}
}
\description{
Stop a running a Splash server Docker container
}
\note{
you need Docker running on your system and have pulled the container with
      [install_splash()] and started the Splash container with [start_splash()] for this
      to work. You will need the `container` object from [start_splash()] for this to work.
}
\examples{
\dontrun{
install_splash()
splash_container <- start_splash()
stop_splash(splash_container)
}
}
\seealso{
Other splash_docker_helpers: \code{\link{install_splash}},
  \code{\link{start_splash}}
}
\concept{splash_docker_helpers}
