% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-default.R
\name{theme-default}
\alias{theme-default}
\title{The default Splash theme}
\description{
Linear mapping: each 0-9 digit maps to \code{floor(d/9 * 255)}, i.e.
\verb{0, 28, 56, 85, 113, 141, 170, 198, 226, 255}. Named colors are the eight
"popular" colors from TodePond's Splash post.
}
\keyword{internal}
