% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-data-frame-har.R
\name{tidy_har}
\alias{tidy_har}
\title{Turn a gnHARly HAR object into a tidy data frame (tibble)}
\usage{
tidy_har(.x, include_content = TRUE)
}
\arguments{
\item{include_content}{if \code{TRUE} (the default) the encoded element content will be returned in the data frame}

\item{x}{A \code{harentry} object}
}
\value{
data frame (tibble)
}
\description{
Turn a gnHARly HAR object into a tidy data frame (tibble)
}
