% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-data-frame-har.R
\name{as_data_frame.harentry}
\alias{as_data_frame.harentry}
\alias{as_data_frame.harentries}
\alias{as_data_frame.har}
\alias{as.data.frame.har}
\alias{as.data.frame.harentries}
\alias{as.data.frame.harentry}
\title{Turns a "HAR"-like object into a data frame(tibble)}
\usage{
\method{as_data_frame}{harentry}(harentry_obj)

\method{as_data_frame}{harentries}(harentries_obj)

\method{as_data_frame}{har}(har_obj)

\method{as.data.frame}{har}(har_obj)

\method{as.data.frame}{harentries}(harentries_obj)

\method{as.data.frame}{harentry}(harentry_obj)
}
\arguments{
\item{harentry_obj}{A \code{harentry} object}

\item{harentries_obj}{A \code{harentry} object}

\item{har_obj}{A \code{har} object}
}
\value{
data frame (tibble)
}
\description{
Turns a "HAR"-like object into a data frame(tibble)
}
