% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{get_content_size}
\alias{get_content_size}
\alias{get_body_size}
\alias{get_headers_size}
\title{Retrieve size of content | body | headers}
\usage{
get_content_size(har_resp_obj)

get_body_size(har_resp_obj)

get_headers_size(har_resp_obj)
}
\arguments{
\item{har_resp_obj}{HAR response object}
}
\description{
Retrieve size of content | body | headers
}
\seealso{
Other splash_har_helpers: \code{\link{get_content_type}},
  \code{\link{get_har_entry}},
  \code{\link{get_header_val}}, \code{\link{get_headers}},
  \code{\link{get_request_type}},
  \code{\link{get_request_url}},
  \code{\link{get_response_body}},
  \code{\link{get_response_url}},
  \code{\link{har_entry_count}}
}
\concept{splash_har_helpers}
