% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get_response_body}
\alias{get_response_body}
\title{Retrieve the body content of a HAR entry}
\usage{
get_response_body(har_resp_obj, type = c("raw", "text"))
}
\arguments{
\item{har_resp_obj}{HAR response object}

\item{type}{return type. If \code{raw} (default) then a raw vector of the content is returned.
If \code{text} then a character vector.}
}
\value{
A \code{raw} vector of the content or \code{NULL} or a \code{character} if \code{type} == \code{text}
}
\description{
Retrieve the body content of a HAR entry
}
\seealso{
Other splash_har_helpers: \code{\link{get_content_size}},
  \code{\link{get_content_type}},
  \code{\link{get_har_entry}},
  \code{\link{get_request_type}},
  \code{\link{get_request_url}},
  \code{\link{har_entry_count}}
}
