% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{har_entry_count}
\alias{har_entry_count}
\title{Retrieves number of HAR entries in a response}
\usage{
har_entry_count(x)
}
\arguments{
\item{x}{can be a `har` object, `harlog` object or `harentries` object}
}
\description{
Retrieves number of HAR entries in a response
}
