% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splashr.r
\docType{data}
\name{splash}
\alias{splash}
\alias{splash_local}
\title{Configure parameters for connecting to a Splash server}
\format{An object of class \code{list} of length 2.}
\usage{
splash(host, port = 8050L)

splash_local
}
\arguments{
\item{host}{host or IP address}

\item{port}{port the server is running on (default is 8050)}
}
\description{
Configure parameters for connecting to a Splash server
}
\keyword{datasets}
