% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splashr.r
\name{splash}
\alias{splash}
\title{Configure parameters for connecting to a Splash server}
\usage{
splash(host, port = 8050L)
}
\arguments{
\item{host}{host or IP address}

\item{port}{port the server is running on (default is 8050)}
}
\description{
Configure parameters for connecting to a Splash server
}
