% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splashr.R
\docType{data}
\name{splash}
\alias{splash}
\alias{splash_local}
\title{Configure parameters for connecting to a Splash server}
\format{An object of class \code{list} of length 4.}
\usage{
splash(host, port = 8050L, user = NULL, pass = NULL)

splash_local
}
\arguments{
\item{host}{host or IP address}

\item{port}{port the server is running on (default is 8050)}

\item{user, pass}{leave `NULL` if basic auth is not configured. Otherwise,
fill in what you need for basic authentication.}
}
\description{
Configure parameters for connecting to a Splash server
}
\examples{
\dontrun{
sp <- splash()
}
}
\keyword{datasets}
