% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splashr.r
\name{splash_active}
\alias{splash_active}
\title{Test if a Splash server is up}
\usage{
splash_active(splash_obj)
}
\arguments{
\item{splash_obj}{A splash connection object}
}
\value{
\code{TRUE} if Slash server is running, otherwise \code{FALSE}
}
\description{
Test if a Splash server is up
}
