% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{splash_click}
\alias{splash_click}
\title{Trigger mouse click event in web page.}
\usage{
splash_click(splash_obj, x, y)
}
\arguments{
\item{splash_obj}{splashr object}

\item{x, y}{coordinates (distances from the left or top, relative to the current viewport)}
}
\description{
Trigger mouse click event in web page.
}
\seealso{
Other splash_dsl_functions: \code{\link{splash_add_lua}},
  \code{\link{splash_focus}}, \code{\link{splash_go}},
  \code{\link{splash_har_reset}}, \code{\link{splash_har}},
  \code{\link{splash_html}}, \code{\link{splash_png}},
  \code{\link{splash_press}}, \code{\link{splash_release}},
  \code{\link{splash_send_keys}},
  \code{\link{splash_send_text}}, \code{\link{splash_wait}}
}
\concept{splash_dsl_functions}
