% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splashr.r
\name{splash_debug}
\alias{splash_debug}
\alias{print.splash_debug}
\title{Retrieve debug-level info for a Splash server}
\usage{
splash_debug(splash_obj = splash_local)

\method{print}{splash_debug}(x, ...)
}
\arguments{
\item{splash_obj}{A splash connection object}
}
\description{
Retrieve debug-level info for a Splash server
}
\examples{
\dontrun{
sp <- splash()
splash_debug(sp)
}
}
\seealso{
Other splash_info_functions: \code{\link{splash_active}},
  \code{\link{splash_history}},
  \code{\link{splash_perf_stats}},
  \code{\link{splash_version}}
}
\concept{splash_info_functions}
\keyword{internal}
