% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.r
\name{splash_focus}
\alias{splash_focus}
\title{Focus on a document element provided by a CSS selector}
\usage{
splash_focus(splash_obj, selector)
}
\arguments{
\item{splash_obj}{splashr object}

\item{selector}{valid CSS selector}
}
\description{
Focus on a document element provided by a CSS selector
}
\references{
See \href{https://splash.readthedocs.io/en/stable/scripting-ref.html#splash-send-text}{the docs} for more info
}
