% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.r
\name{splash_go}
\alias{splash_go}
\title{Go to an URL.}
\usage{
splash_go(splash_obj, url)
}
\arguments{
\item{splash_obj}{splashr object}

\item{url}{- URL to load;}
}
\description{
This is similar to entering an URL in a browser address bar, pressing Enter and waiting
until page loads.
}
\examples{
\dontrun{
splash_local \%>\%
  splash_response_body(TRUE) \%>\%
  splash_user_agent(ua_macos_chrome) \%>\%
  splash_go("https://rud.is/b") \%>\%
  splash_wait(2) \%>\%
  splash_har() -> rud_har
}
}
\seealso{
Other splash_dsl_functions: \code{\link{splash_add_lua}},
  \code{\link{splash_click}}, \code{\link{splash_focus}},
  \code{\link{splash_har_reset}}, \code{\link{splash_har}},
  \code{\link{splash_html}}, \code{\link{splash_png}},
  \code{\link{splash_press}}, \code{\link{splash_release}},
  \code{\link{splash_send_keys}},
  \code{\link{splash_send_text}}, \code{\link{splash_wait}}
}
