% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.r
\name{splash_har_reset}
\alias{splash_har_reset}
\title{Drops all internally stored HAR records.}
\usage{
splash_har_reset(splash_obj, keys)
}
\arguments{
\item{splash_obj}{splashr object}
}
\description{
Drops all internally stored HAR records.
}
