% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splashr.r
\name{splash_history}
\alias{splash_history}
\title{Get information about requests/responses for the pages loaded}
\usage{
splash_history(splash_obj)
}
\arguments{
\item{splash_obj}{A splash connection object}
}
\description{
Get information about requests/responses for the pages loaded
}
