% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.r
\name{splash_images}
\alias{splash_images}
\title{Enable/disable images}
\usage{
splash_images(splash_obj, enable = TRUE)
}
\arguments{
\item{splash_obj}{splashr object}

\item{enable}{logical}
}
\description{
By default, images are enabled. Disabling of the images can save a lot of network
traffic (usually around ~50%) and make rendering faster. Note that this option can
affect the JavaScript code inside page: disabling of the images may change sizes and
positions of DOM elements, and scripts may read and use them.
}
\examples{
\dontrun{
splash_local \%>\%
  splash_images(TRUE) \%>\%
  splash_user_agent(ua_macos_chrome) \%>\%
  splash_go("https://rud.is/b") \%>\%
  splash_wait(2) \%>\%
  splash_har() -> rud_har
}
}
\seealso{
Other splash_dsl_attributes: \code{\link{splash_enable_javascript}},
  \code{\link{splash_plugins}},
  \code{\link{splash_private_mode}},
  \code{\link{splash_response_body}}
}
\concept{splash_dsl_attributes}
