% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splashr.r
\name{splash_perf_stats}
\alias{splash_perf_stats}
\title{Get Splash performance-related statistics}
\usage{
splash_perf_stats(splash_obj)
}
\arguments{
\item{splash_obj}{A splash connection object}
}
\description{
Get Splash performance-related statistics
}
