% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.r
\name{splash_png}
\alias{splash_png}
\title{Return a screenshot of a current page in PNG format.}
\usage{
splash_png(splash_obj)
}
\arguments{
\item{splash_obj}{splashr object}
}
\value{
a \link{magick} image object
}
\description{
Similar to \link{render_png} but used in a script context. Should be the LAST element in
a DSL script chain as this will execute the script and return the PNG content
}
\examples{
\dontrun{
splash_local \%>\%
  splash_go("https://rud.is/b") \%>\%
  splash_wait(2) \%>\%
  splash_png()
}
}
