% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splashr.r
\name{splash_version}
\alias{splash_version}
\title{Get Splash version information}
\usage{
splash_version(splash_obj)
}
\arguments{
\item{splash_obj}{A splash connection object}
}
\description{
Get Splash version information
}
