% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{splash_wait}
\alias{splash_wait}
\title{Wait for a period time}
\usage{
splash_wait(splash_obj, time = 2)
}
\arguments{
\item{splash_obj}{splashr object}

\item{time}{number of seconds to wait}
}
\description{
When script is waiting WebKit continues processing the webpage
}
\examples{
\dontrun{
splash_local \%>\%
  splash_response_body(TRUE) \%>\%
  splash_user_agent(ua_macos_chrome) \%>\%
  splash_go("https://rud.is/b") \%>\%
  splash_wait(2) \%>\%
  splash_har() -> rud_har
}
}
\seealso{
Other splash_dsl_functions: \code{\link{splash_add_lua}},
  \code{\link{splash_click}}, \code{\link{splash_focus}},
  \code{\link{splash_go}}, \code{\link{splash_har_reset}},
  \code{\link{splash_har}}, \code{\link{splash_html}},
  \code{\link{splash_png}}, \code{\link{splash_press}},
  \code{\link{splash_release}},
  \code{\link{splash_send_keys}},
  \code{\link{splash_send_text}}
}
\concept{splash_dsl_functions}
