% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splashr-package.R
\docType{package}
\name{splashr}
\alias{splashr}
\title{Tools to Work with the 'Splash' JavaScript Rendering Service}
\description{
'Splash' \url{https://github.com/scrapinghub/splash} is a 'JavaScript' rendering service.
It’s a lightweight web browser with an 'HTTP' API, implemented in 'Python' using 'Twisted'
and 'QT' and provides some of the core functionality of the 'RSelenium' or 'seleniumPipes'
R pacakges in a lightweight footprint.
}
\details{
Some of 'Splash' features include the ability to process
multiple webpages in parallel; retrieving 'HTML' results and/or take screenshots; disabling
images or use 'Adblock Plus' rules to make rendering faster; executing custom 'JavaScript' in
page context; getting detailed rendering info in 'HAR' format.
}
\author{
Bob Rudis (bob@rud.is)
}
