% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.r
\name{start_splash}
\alias{start_splash}
\title{Start a Splash server Docker container}
\usage{
start_splash()
}
\value{
`harbor` `container` object
}
\description{
Start a Splash server Docker container
}
\note{
you need Docker running on your system and have pulled the container with
      [install_spash] for this to work. You should save the resultant `host`
      object for use in [stop_splash].
}
\examples{
\dontrun{
install_splash()
splash_container <- start_splash()
stop_splash(splash_container)
}
}
