% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker-splash.r
\name{start_splash}
\alias{start_splash}
\title{Start a Splash server Docker container}
\usage{
start_splash(host = harbor::localhost, add_tempdir = TRUE)
}
\arguments{
\item{host}{Docker host; defauolts to `localhost`}

\item{add_tempdir}{if `TRUE` then a local temporary directory (made with [tempdir]())
will be added to the mount configuration for use with [render_file](). You will need to
    ensure the necessary system temp dirs are accessible as a mounts. For
    macOS this means adding `/private` to said Docker config.}
}
\value{
`harbor` `container` object
}
\description{
Start a Splash server Docker container
}
\note{
you need Docker running on your system and have pulled the container with
      [install_spash] for this to work. You should save the resultant `host`
      object for use in [stop_splash].
}
\examples{
\dontrun{
install_splash()
splash_container <- start_splash()
stop_splash(splash_container)
}
}
