% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker-splash.r, R/docker.r
\name{start_splash}
\alias{start_splash}
\alias{start_splash}
\title{Start a Splash server Docker container}
\usage{
start_splash(host = harbor::localhost)

start_splash(host = harbor::localhost)
}
\arguments{
\item{host}{Docker host; defauolts to `localhost`}

\item{add_tempdir}{This is `FALSE` initially since you could try to run
the splash image on a remote system. It has to be a local one for this to work.
If `TRUE` then a local temporary directory (made with [tempdir]())
will be added to the mount configuration for use with [render_file](). You will need to
    ensure the necessary system temp dirs are accessible as a mounts. For
    macOS this means adding `/private` to said Docker config.}

\item{host}{Docker host; defauolts to `localhost`}
}
\value{
`harbor` `container` object

`harbor` `container` object
}
\description{
If using this in an automation context, you should consider adding a
`Sys.sleep(3)` (or higher) after starting the docker container.

Start a Splash server Docker container
}
\note{
you need Docker running on your system and have pulled the container with
      [install_spash] for this to work. You should save the resultant `host`
      object for use in [stop_splash].

you need Docker running on your system and have pulled the container with
      [install_spash] for this to work. You should save the resultant `host`
      object for use in [stop_splash].
}
\examples{
\dontrun{
install_splash()
splash_container <- start_splash()
stop_splash(splash_container)
}
\dontrun{
install_splash()
splash_container <- start_splash()
stop_splash(splash_container)
}
}
